Rename widget in preparation for possible second contract widget
