+++ app/src/main/res/values/strings.xml
@@ -3,5 +3,5 @@
     <string name="mission_widget_normal">Missions Normal View</string>
     <string name="mission_widget_minimal">Missions Minimal View</string>
     <string name="mission_widget_large">Mission Large View</string>
-    <string name="contract_widget_normal">Contracts Normal View</string>
+    <string name="contract_widget_active">Contracts Active View</string>
 </resources>
